requirements.txt
python-dotenv
langchain_community
langchain_openai
langchain_core
streamlit
langserve
fastapi
uvicorn
sse_starlette
pinecone
streamlit-navigation-bar
boto3
pandas